+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Unreleased
 
-Nothing yet
+* Added support for declaring wildcard audits and trusted entries for crates published using ["Trusted Publishing"](https://crates.io/docs/trusted-publishing) (#671)
 
 # Version 0.10.1 (2025-02-10)